new_requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Mail==0.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
paypalrestsdk==1.13.3
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==25.0.0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
